Remove cropping in preprocess step
Originally I wanted to be able to use the .bmp assets from 1.22
unmodified and do the alpha conversion at runtime, but I
realized that this meant having a hard dependency on lua-vips
and thus LuaJIT, making it impossible to have a web version. So this
is much less than I wanted to do and amounts to bikeshedding
at this point.