Add BMP loading with transparency conversion at runtime
Actually I was wrong. All that we have to do is write a fallback BMP
loader if 5.1 is being used, and use vips if LuaJIT is used instead.
Since BMPs seem to be the primary image format lacking transparency
that is used for elona assets, it seems like this will work for the
time being. We may need to add PNG support later though.