Redesign map generation system (mostly)
This encompasses multiple changes to the map generation system that
have to do with each other:

1. Map generators

The "map_generator" type was redundant. Essentially all it did was
wrap a function that generated a map. There was too much "magic" in
the way of how maps were generated, and it was confusing because there
was nothing stopping you from just writing a function to generate a
map anyway. This change removes "map_generator" and makes all the
logic for the maps they generated plain functions instead.

2. Map templates

Even with the map generation system, for all intents and purposes
you'd usually just use "elona_sys.map_template" if you wanted anything
to do with saving a map. This was used so frequently that there were
dependencies in the base mod for it. So it was moved into the base
code as a first-class feature.

Map templates are needed due to map regeneration, where the logic for
regenerating a map from scratch is separate from simply refreshing it,
because of things like the shop rank of shopkeepers that need to be
preserved. So if you have a map that needs geometry rebuilt in this
way, you can associate a function that will always regenerate a fresh
copy of a specific map each time with an ID, like "vernis", so it can
get regenerated properly. Map generation functions alone don't solve
this because you'd need to call them like generate("vernis"), and
"vernis" isn't defined in a static location anywhere.

3. Global maps

There was this problem that wasn't solved well with the original way
maps were generated.

In vanilla you can say "give me map 'vernis'" and the game will load
Vernis. There is only ever one copy of Vernis, and it is always
generated the same way each time.

However, the map generation system in ON was too flexible, and had no
global storage of "unique" maps. Thus, you had to essentially iterate
all of the map entrances inside the world map (north tyris) in order
to obtain the ID of the map pointing to Vernis. The map for north
tyris had to be loaded entirely in order to do this, which was
wasteful. If this ID were stored in a global lookup table instead,
which says "vernis is map 1", then we could just say
Map.load("vernis"). Worse, you had to do fuzzy comparison to ensure
this map was actually vernis, based on attributes like how the map was
generated, instead of just declaring that it was vernis.

This code change reimplements vanilla's behavior allowing for global
map references. By marking a map template as "unique", this associates
the template's ID with a map UID the first time the map is generated,
so you can do Map.load("elona.vernis") if there is a unique template
with ID "elona.vernis" and things will just work.

(Note: moving between maps is still broken temporarily...)