Implement quick menu
Added an option to KeyHandler to capture the pressed/release state of a
keybind. Updated IKeyInput:run_key_action accordingly to take the
arguments `(pressed, ...)` instead of just `(...)`, where `pressed` is
true or false. This will only be enabled if the user specifically asks
to turn it on, otherwise the behavior is the same.

We should reconsider how to handle diagonal inputs eventually. For
movement commands we have to implement a bunch of ugly logic inside
QuickMenuList to handle the case of selecting a diagonal.