Add open-nefia-context editor thing
Frequently I want to know the following:
- If part of the original code has been ported already.
- Where in the original codebase a region of code in OpenNefia was
  ported from.
- A rough estimate of how much of the original code hasn't been ported
  yet.

Hopefully adding annotations to the existing ported codebase will help
with these wants. Because a lot of the code is split up across many
different places by nature of OpenNefia's event system, it's getting
difficult to keep track of everything.

Also, if there are any bugs, having the locations in the original source
to cross-reference would greatly help tracking down the issue. We might
even be able to add annotations for the other variants, if we're going
to attempt directly porting their features from some version of their
decompiled sources.