Implement basic theming system
At first it will support modifying the base assets. It might be nice as
a general-purpose modding system, but that will remain to be seen.

Essentially how it works is you provide a new entry of `base.theme`, and
add in a set of data entries of the same shape as the target asset,
including `_id` and `_type` on it, and supplying the fields to modify.
These are then merged into shallow copy of the original data. The merged
data has the same schema, so it can be fed into `atlas` and such without
too much fuss.

Future work would allow multiple overrides, allowing the player to
choose from each in a menu, and finally saving their configuration as a
"themepack", which is a configuration dependent on all the themes it
uses.

Performance doesn't seem to be too terrible. The data merging itself
only takes less than a second; it's mostly the building of the asset
canvas that takes the majority of the time when reloading the theme.