Fix bug causing object iterators to become stale
This is because `state.uids` inside the iterators is deepcopied from the
parent, in order to make sure the index of iteration is always
valid. (It's an attempt to prevent the same dilemma with trying to
remove an element from a table while also iterating the same table with
`ipairs`).

However, if an object in the pool gets removed after the iterator is
created, that can cause a nil access, because `state.uids` will not be
updated to reflect the removal. This creation is done in the code inside
`field_logic`, for stepping through each character's turn.