Fix Rand.rnd() not returning large integers
HSP's rand() function will not return numbers exceeding 2^15 (32768).
This means if you try to use `Rand.rnd(n)` as you would intuitively
expect for large numbers, you will get incorrect behavior.

As a compromise, modify Rand.rnd() to use HSP's legacy behavior if `n`
is less than 32768, and otherwise generate a 64-bit value. The
reason we can't replace `Rand.rnd()` and always use 64-bit values is
for compatibility with HSP. Certain routines are dependent on the
random values HSP returns for a specific seed, such as the
generation of random titles. It would also be difficult to discover
other weird edge cases where we want HSP's behavior, so this is the
fix that will be made for now.

This fixes a bug in WeightedSampler (and consequently random
character/item generation) where items of equal weight were not
being sampled in approximately equal probabilities.