Implement IOwned:containing_map()
To do this, implement `IOwned` on `EquipSlots`. This allows you to get
the map that an equipped item is contained in overall, recursing up the
chain of parent locations.

To me this feels weird, because `location` was previously only a field
that got used by `IMapObject`. This is the first time I can remember off
the top of my head that it's being used in a plain class instead of a
map object. But there has to be *some* way of knowing the owner of an
equipped item, and this design is one way of accomplishing this.