More or less port all dungeons
Still some missing things, but you can now enter dungeons and expect to
be able to go between floors.

Here's how the generation system works so far:

1. Create an area template's `on_generate_floor` callback.
1. Use a function in `DungeonTemplate` to get back a dungeon generator
and parameter list.
2. Pass the two results along with the area and floor number into
`DungeonMap.generate()`.
3. Set the archetype and any properties, then return the finished map
from the callback.

This way, the randomly selected generator and parameters are decoupled
from the generation process, in case you want to further modify how the
dungeon is generated before it gets finalized. It should be reusable
enough for other mods to pick up.