Add advice system
Let's try this out.

Adds an advice system similar in functionality to that of Emacs. It
allows for adding new code to functions that are exposed as part of
modules. This would be a really easy way to change the game's numerous
formulas without needing to add a new event type for every single one.

Using this, I was able to change the sprite of every generated character
and set the material of all generated items to Rubynus.

I'm wondering how stable this will be, though. The system depends on the
ability for mods to change the contents of code exposed in modules by
directly editing the module's table. I think I will allow this in the
meantime with the assumption that nobody will do something stupid like
replace an advice-generated function with something else (since the
advice system uses the function it generates on the module table as a
key to the table of active advice internally). I'll save users from
themselves when there are actually any users to save.

Also, first commit on my new Kinesis. It's a pain in the butt to get
used to, but no more wrist strain! Woot!