Implement a poor man's matplotlib
I always wanted this, for some reason. I just wanted to be able to make
line plots for things like RAM usage without having to use an external
program. Maybe it could be useful for tracking game stats in the future?
Not sure. Regardless, it's in a mod, so it could be moved out if need
be.