Visual AI - small cleanup, remove readme
I typically want to write the README first to imagine how I'd use the
thing when it's finally finished.

Issue now is that the bugfixes in the engine are tied to the mod code,
which makes the commit history a bit unclean. From now on the
nice-to-have mods should get their own git repos to separate out the
two...