Add :iter_merged_enchantments() to ICharaEquip -> IChara
The list of total powers for each enchantment is cached, and the
cache is cleared on save (TODO) or when the enchantment list is modified
using the API.

Only issue is what happens if the API isn't used.

But this way we are able to both iterate all enchantments, and also know
the breakdown of where each enchantment gets its power.