Change how location is set for map objects
Now it will only ever be set by the `pool` class. Everyone else will
traverse up either a map object's `location` or a location's
`_parent` to find the parent location.

Manually setting `location` was too confusing, because the object
was not actually "owned" by any data structure in the parent
location that did this.

It caused some nasty bugs while I was testing though, so I'm wary of
regressions.