Fix not setting map object location before running finalize
Some of the finalize hooks rely on calling :current_map() or similar.
The change we just made causes the `location` field to no longer get
set on the object by hand when calling `Item.create()` and similar,
so now this started to fail when :current_map() started returning
nil. However, this is a very good thing because previously, the
location hadn't actually put the object into its pool yet - it had
only set the `location` field, so technically it was not the true
location of the object. So that had revealed a bug that was lying
hidden the whole time.

This change ensures the location is set before running :finalize()
and related hooks on the object. Now the object will be properly
owned by a location before the hooks are run.

Also, this revealed another bug with opening an item if it's in your
inventory instead of on the ground. The item's position will not be
updated if it's inside your inventory when you move. However, the
code to generate the items from the opened chest still used those
outdated coordinates on the item itself. Now that the ownership is
actually correct we can use :containing_map() to get the position of
the thing that owns the item, to determine the actual position to
place the items on the map.