Make location field on map objects immutable
This is somewhat more complicated than the rest, so I'm expecting
regressions.

location gets stored on the metatable. Now that the metatable is keeping
individual state per object, it no longer makes sense to share them
between cloned objects, so a bit of code was added to
MapObject.clone() to account for that.