Add "staying characters" thing
This is to emulate a specific feature in vanilla: the worker system.

When you assign a shopkeeper, the character's "current area" gets set
to that of the current map. When you leave maps, the game checks if
any characters on the map have a "current area" set and different
than the current one, and hides them by setting their state to
"AllyWork".

The implementation here is slightly different. Instead the staying
characters get moved into a global StayingCharas container when the map
is changed, and restored afterwards.