Support card/figure rendering in world map and in menus
This requires a slight change to some of the batches so things like the
character in the figure can be rendered properly. The new MapObjectBatch
will properly draw the custom drawable things as well as regular chips.

The one thing that this doesn't handle is the throwing animation -
if a figure is thrown, there will be no image on it until it lands. I
think this is a good compromise for now.