Fix cloning of EquipSlots
What happens when a map object is cloned is: any map objects that
are found recursively are assigned new UIDs. But, of course, if you
use those UIDs in the parent object for bookkeeping they'll be out
of date if the object gets cloned.

To solve this, add a new interface (ICloneable) which lets you
insert some logic for updating the UIDs when MapObject.clone() gets
called.