Change on_start() dialog callback to run before text() callbacks
This is to fix a dependency on on_start() that text() previously had
when turning in quests. It became a problem when updating the mage
guild's quota, because the quota update happened after the text was
localized from I18N.get(), so the quota amount was printed as 0 in
the dialog.

To solve the quest issue, separate the quest state "completed" into
"completed" and "finished", and clean up finished quests when
generating new ones.