Merge StayingCharas and StayingAdventurers
The reality is that areas are very lightweight, while maps are very
heavyweight. To get a map's information you have to load it from
disk. This meant that when attaching a character to a map, the map had
to be loaded first. This became a problem when "spawning"
adventurers inside a random area, as with the old implementation of
StayingCharas you had no choice but to generate/load the area's
starting map to attach the adventurer there, which was very expensive.

Now we can simply associate a character with a map and floor, and
move them there whenever the floor first gets generated/serialized.