Transfer staying characters in both previous and next map
Before the staying characters logic only operated on characters in the
current map, because it was only being used for allies, which had
already been moved to the new map by the time the transfer happened. But
for adventurers, which are left behind in the previous map, the
transfer was not occurring.