Fix binding of events when first building objects
The problem is we want to be able to have handlers for
"base.on_object_finalized", but before the events were only being
bound when the object was instantiated, after the finalize stage. Those
events shouldn't be in "base.on_object_instantiated" because they get
run every time the object is loaded, not just on creation. But we also
need to bind the events on instantiate when the object is loaded.

To fix this, add the flag "no_bind_events" to IObject:instantiate() and
set this to true only when Object.finalize() is called. The events will
be bound inside IObject:finalize() in that case. In every other case
they will be bound in IObject:instantiate().