Update various events to use "on_map_entered" instead of "on_map_changed"
This is the problem I was wanting to fix. "on_map_changed" gets fired
both when the map is changed *and* when the map is loaded.
"on_map_entered" in my mind was supposed to mean "when the player
transitions from one map to another." The map init code will not run
"on_map_entered" events if loading a save, so this should work (?).