Rework how god offerings are calculated
Instead of needing to tag the item's prototype with the gods the item is
offerable to, you tag it with an item category. Precious items will
always be excluded. This should make it easier to just add new items
and have the offerings automatically be compatible with existing
gods.

Some of the offerings in vanilla don't belong uniformly to one category.
For example, there is no category in vanilla that corresponds to the
exact set of items offerable to Mani or Kumiromi, so some new
categories just for those items were created.