[ffhp] Make mapping identifiers more general
The interactions between systems are getting more complicated. It seems
like we'd need to be able to combine the theming system and
base.data_ext in order to implement ffhp.item_ex_mapping in an
"automatic" way, so that the FFHP mappings don't need to be created
by hand for every new custom item.