Advanced drawing APIs
These are for drawing things above and below the game canvas. Elona
wasn't designed for high-DPI screens, so I thought that maybe it would
be useful to be able to put some information in some fringes around the
screen. I saw a video for the rerelease of a Shiren the Wanderer title
that did this.

These kinds of features give me some anxiety, as they involve balancing
out the fact that this is currently an experimental project to poke
around at whatever ideas I have about building a living system with the
fact that this might ultimately become useful to people, meaning that
restraint in piling on features would eventually be required. This
change was easy enough to add because the drawing system wasn't terribly
complicated, but each additional feature adds even more complexity. Is
the utility of this feature ultimately worth the maintenance burden? It
seems that only time will be able to tell.

As for why I wanted these features in the first place, the original idea
was that I wanted to create a mod where Yuzuki Yukari becomes a backseat
gamer and commentates on how terribly I'm playing Elona in real-time.